MetricMeasurements are now recorded as doubles rather than floats. A getRoundedResult() method was implemented in the MetricMeasurement class that is able to trim the size of the double down to a specified number of decimal place digits. The "hot" and "cold" thresholds defined in metrics.xsd are now parsed in as doubles, also. The ClassPropertiesPanel and ClassPropertiesTableModel, as well as the visualisations that display metric readings as part of their geometric representations of classes, display these values with 5 fractional digits proceeding the point.

+++ src/org/lukep/javavis/metrics/MetricAttribute.java
@@ -16,8 +16,8 @@ public class MetricAttribute {
 	protected List<String> appliesTo;
 	protected Class<IMeasurableVisitor> visitor;
 	protected String argument;
-	protected float cold;
-	protected float hot;
+	protected double cold;
+	protected double hot;
 	
 	public MetricAttribute(String name, String nameInternal, MetricType type, List<String> appliesTo) {
 		this.name = name;
@@ -78,11 +78,11 @@ public class MetricAttribute {
 		return argument;
 	}
 
-	public float getCold() {
+	public double getCold() {
 		return cold;
 	}
 
-	public float getHot() {
+	public double getHot() {
 		return hot;
 	}
