+++ src/org/lukep/javavis/visualisation/views/CircleView.java
@@ -43,7 +43,7 @@ public class CircleView extends AbstractVisualisationView {
 		graph.addColumn("type", String.class);
 		graph.addColumn("name", String.class);
 		graph.addColumn("model", IGenericModelNode.class);
-		graph.addColumn("metricMeasurement", float.class);
+		graph.addColumn("metricMeasurement", double.class);
 		
 		// create package vertices
 		ProjectModel modelStore = wspContext.getModelStore();
@@ -67,7 +67,7 @@ public class CircleView extends AbstractVisualisationView {
 			// grab metric measurement if applicable
 			if (model instanceof IMeasurableNode
 					&& wspContext.getMetric().testAppliesTo(model.getModelTypeName())) {
-				curNode.setFloat("metricMeasurement", 
+				curNode.setDouble("metricMeasurement", 
 						((IMeasurableNode)(model)).getMetricMeasurement(
 								wspContext.getMetric()).getResult());
 			}
