+++ src/org/lukep/javavis/visualisation/views/ClassHierarchyView.java
@@ -140,7 +140,7 @@ public class ClassHierarchyView extends AbstractVisualisationView {
 		graph.addColumn("type", String.class);
 		graph.addColumn("name", String.class);
 		graph.addColumn("model", IGenericModelNode.class);
-		graph.addColumn("metricMeasurement", float.class);
+		graph.addColumn("metricMeasurement", double.class);
 		
 		// create package vertices
 		ProjectModel modelStore = wspContext.getModelStore();
@@ -199,7 +199,7 @@ public class ClassHierarchyView extends AbstractVisualisationView {
 			// grab metric measurement if applicable
 			if (model instanceof IMeasurableNode
 					&& wspContext.getMetric().testAppliesTo(model.getModelTypeName())) {
-				curNode.setFloat("metricMeasurement", 
+				curNode.setDouble("metricMeasurement", 
 						((IMeasurableNode)(model)).getMetricMeasurement(
 								wspContext.getMetric()).getResult());
 			}
